+++ src/types.ts
@@ -13,6 +13,8 @@ export interface GlobalOptions {
   packageManager?: PackageManager;
   monorepo?: boolean;
   fix?: boolean;
+  fixCompat?: boolean;
+  skipSharedCheck?: boolean;
 }
 
 export interface WriteConfigsOptions {
@@ -47,6 +49,8 @@ export interface UpdatePackageResult {
   addedDependencies: string[];
   addedScripts: string[];
   updatedDependencies: string[];
+  addedOverrides: string[];
+  updatedOverrides: DepVersionChange[];
   updatedScripts: string[];
   wroteFile: boolean;
 }
@@ -98,6 +102,8 @@ export interface UpdatePackageForUpdateResult {
   addedDependencies: string[];
   addedScripts: string[];
   updatedDependencies: DepVersionChange[];
+  addedOverrides: string[];
+  updatedOverrides: DepVersionChange[];
   updatedScripts: string[];
   wroteFile: boolean;
 }